Portrait Puzzle Proof of Concept
I have it to where the portraits will switch materials when the chess piece is "throwin into the fire". I just did this with fake models that I assigned similar values to  what would be in game. Also made a snapzone on a standin gramaphone and made sure it would play audio when snapped in.